Extract site REST proxy, media MIME, and installed-app detection to @studio/common
Three small, independent shared libs, each rewiring one desktop spot to delegate: the WordPress REST proxy (wordpress-rest-api.ts), local-media MIME lookup (ipc-handlers.ts), and installed editor/terminal detection (is-installed.ts, now a re-export; its test moves to @studio/common). Behavior is unchanged except one hardening that comes with the shared REST proxy: its request.path branch now validates the resolved URL stays within the site's own REST root, closing a latent SSRF. Part of splitting the studio ui PoC (#3953).
