+++ apps/studio/src/lib/is-installed.ts
@@ -1,180 +1,9 @@
-import { app } from 'electron';
-import fs from 'fs';
-import path from 'path';
-import { findOnPath } from 'src/lib/find-on-path';
-import { SUPPORTED_EDITORS, supportedEditorConfig } from 'src/modules/user-settings/lib/editor';
-import { SUPPORTED_TERMINALS, terminalConfig } from 'src/modules/user-settings/lib/terminal';
-
-type PlatformPaths = {
-	[ K in keyof InstalledApps ]: string[];
-};
-
-function getProgramFilesPath(): string {
-	if ( process.platform !== 'win32' ) {
-		return 'C:\\Program Files';
-	}
-
-	// This env var dinamically points to the Program Files path
-	// See https://stackoverflow.com/a/9608782
-	const programFiles = process.env.ProgramFiles;
-	if ( programFiles ) {
-		return programFiles;
-	}
-
-	// Fallback to default path if environment variable is not available
-	return 'C:\\Program Files';
-}
-
-function getLocalProgramsPath(): string {
-	if ( process.platform !== 'win32' ) {
-		return app.getPath( 'appData' );
-	}
-
-	const localAppData = process.env.LOCALAPPDATA;
-	if ( localAppData ) {
-		return path.win32.join( localAppData, 'Programs' );
-	}
-
-	// Fallback to electron's appData path if environment variable is not available
-	return path.win32.join( app.getPath( 'appData' ), 'Local', 'Programs' );
-}
-
-// Define installation paths for each IDE by platform
-const installationPaths: Record< string, PlatformPaths > = {
-	darwin: {
-		antigravity: [ 'Antigravity.app' ],
-		vscode: [ 'Visual Studio Code.app' ],
-		phpstorm: [ 'PhpStorm.app' ],
-		cursor: [ 'Cursor.app' ],
-		windsurf: [ 'Windsurf.app' ],
-		webstorm: [ 'WebStorm.app' ],
-		sublime: [ 'Sublime Text.app' ],
-		zed: [ 'Zed.app' ],
-		iterm: [ 'iTerm.app' ],
-		terminal: [ 'Terminal.app' ],
-		warp: [ 'Warp.app' ],
-		ghostty: [ 'Ghostty.app' ],
-	},
-	// Linux editor and terminal detection is driven by `linuxCommands`
-	// (resolved via $PATH in `isInstalled`), so those entries are intentionally
-	// omitted from this map. Only the non-resolvable iterm entry lives here.
-	linux: {
-		antigravity: [],
-		vscode: [],
-		phpstorm: [],
-		cursor: [],
-		windsurf: [],
-		webstorm: [],
-		sublime: [],
-		zed: [],
-		iterm: [],
-		terminal: [],
-		warp: [],
-		ghostty: [],
-	},
-	win32: {
-		antigravity: [
-			path.win32.join( getLocalProgramsPath(), 'Antigravity' ),
-			path.win32.join( getProgramFilesPath(), 'Google\\Antigravity' ),
-		],
-		vscode: [
-			path.win32.join( getProgramFilesPath(), 'Microsoft VS Code' ),
-			path.win32.join( getLocalProgramsPath(), 'Microsoft VS Code' ),
-		],
-		phpstorm: [
-			path.win32.join( getProgramFilesPath(), 'JetBrains\\PhpStorm' ),
-			path.win32.join( getLocalProgramsPath(), 'PhpStorm' ),
-		],
-		cursor: [
-			path.win32.join( getProgramFilesPath(), 'Cursor' ),
-			path.win32.join( getLocalProgramsPath(), 'cursor' ),
-		],
-		windsurf: [
-			path.win32.join( getProgramFilesPath(), 'Windsurf' ),
-			path.win32.join( getLocalProgramsPath(), 'Windsurf' ),
-		],
-		webstorm: [
-			path.win32.join( getProgramFilesPath(), 'JetBrains\\WebStorm' ),
-			path.win32.join( getLocalProgramsPath(), 'WebStorm' ),
-		],
-		sublime: [
-			path.win32.join( getProgramFilesPath(), 'Sublime Text' ),
-			path.win32.join( getProgramFilesPath(), 'Sublime Text 4' ),
-			path.win32.join( getProgramFilesPath(), 'Sublime Text 3' ),
-		],
-		zed: [ path.win32.join( getLocalProgramsPath(), 'Zed' ) ],
-		iterm: [],
-		terminal: [],
-		warp: [
-			path.win32.join( getLocalProgramsPath(), 'Warp' ),
-			path.win32.join( getProgramFilesPath(), 'Warp' ),
-		],
-		ghostty: [],
-	},
-};
-
-if ( process.platform === 'darwin' ) {
-	const systemApplications = '/Applications';
-	const userApplications = path.join( app.getPath( 'home' ), 'Applications' );
-
-	Object.keys( installationPaths.darwin ).forEach( ( ide ) => {
-		const appName = installationPaths.darwin[ ide as keyof InstalledApps ][ 0 ];
-		if ( appName ) {
-			installationPaths.darwin[ ide as keyof InstalledApps ] = [
-				path.join( systemApplications, appName ),
-				path.join( userApplications, appName ),
-			];
-		}
-	} );
-} else if ( process.platform === 'win32' ) {
-	// For JetBrains IDEs, check for version-specific folders
-	[ 'phpstorm', 'webstorm' ].forEach( ( ide ) => {
-		const basePaths = installationPaths.win32[ ide as keyof InstalledApps ];
-		const jetbrainsDir = path.win32.join( getProgramFilesPath(), 'JetBrains' );
-
-		if ( fs.existsSync( jetbrainsDir ) ) {
-			const entries = fs.readdirSync( jetbrainsDir );
-
-			entries.forEach( ( entry ) => {
-				if ( entry.toLowerCase().includes( ide ) ) {
-					basePaths.push( path.win32.join( jetbrainsDir, entry ) );
-				}
-			} );
-		}
-	} );
-}
-
-function isSupportedEditor(
-	key: keyof InstalledApps
-): key is ( typeof SUPPORTED_EDITORS )[ number ] {
-	return ( SUPPORTED_EDITORS as readonly string[] ).includes( key );
-}
-
-function isSupportedTerminal(
-	key: keyof InstalledApps
-): key is ( typeof SUPPORTED_TERMINALS )[ number ] {
-	return ( SUPPORTED_TERMINALS as readonly string[] ).includes( key );
-}
-
-export function isInstalled( key: keyof InstalledApps ): boolean {
-	const platform = process.platform;
-
-	// On Linux, editors and terminals live in many places (apt, snap, flatpak,
-	// ~/.local/bin, …). Resolve them against $PATH using their `linuxCommands`
-	// instead of hardcoded paths.
-	if ( platform === 'linux' ) {
-		if ( isSupportedEditor( key ) ) {
-			const editor = supportedEditorConfig[ key ];
-			return editor.linuxCommands.some( ( command ) => findOnPath( command ) !== null );
-		}
-		if ( isSupportedTerminal( key ) ) {
-			const terminal = terminalConfig[ key ];
-			return terminal.linuxCommands.some( ( command ) => findOnPath( command ) !== null );
-		}
-	}
-
-	const paths = installationPaths[ platform ]?.[ key ];
-
-	// Return true if any of the possible paths exist
-	return paths.some( ( pathStr: string ) => pathStr && fs.existsSync( pathStr ) );
-}
+// Installed-app detection now lives in @studio/common so the desktop app and
+// the local web server (`studio ui`) share one implementation. This re-export
+// keeps existing desktop imports — and the tests that mock this module — working.
+export {
+	detectInstalledApps,
+	isInstalled,
+	type AppKey,
+	type InstalledApps,
+} from '@studio/common/lib/user-settings/installed-apps';
